Using DatabaseHandler for properties as one handler per object
