+++ src/Microsoft.SqlTools.ServiceLayer/ObjectManagement/ObjectTypes/Database/DatabaseHandler.cs
@@ -85,7 +85,7 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
         public override Task<InitializeViewResult> InitializeObjectView(InitializeViewRequestParams requestParams)
         {
             // create a default data context and database object
-            using (var dataContainer = CreateDatabaseDataContainer(requestParams.ConnectionUri, ConfigAction.Create))
+            using (var dataContainer = CreateDatabaseDataContainer(requestParams, ConfigAction.Create))
             {
                 if (dataContainer.Server == null)
                 {
@@ -104,6 +104,27 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
                         {
                             ObjectInfo = new DatabaseInfo()
                         };
+
+                        // Collect the Database properties information
+                        if (!requestParams.IsNewObject)
+                        {
+                            var smoDatabaseProperties = dataContainer.SqlDialogSubject as Database;
+                            databaseViewInfo.ObjectInfo = new DatabaseInfo()
+                            {
+                                Name = smoDatabaseProperties.Name,
+                                CollationName = smoDatabaseProperties.Collation,
+                                DateCreated = smoDatabaseProperties.CreateDate.ToString(),
+                                LastDatabaseBackup = smoDatabaseProperties.LastBackupDate == DateTime.MinValue ? "None" : smoDatabaseProperties.LastBackupDate.ToString(),
+                                LastDatabaseLogBackup = smoDatabaseProperties.LastLogBackupDate == DateTime.MinValue ? "None" : smoDatabaseProperties.LastLogBackupDate.ToString(),
+                                MemoryAllocatedToMemoryOptimizedObjectsInMb = ConvertKbtoMbString(smoDatabaseProperties.MemoryAllocatedToMemoryOptimizedObjectsInKB),
+                                MemoryUsedByMemoryOptimizedObjectsInMb = ConvertKbtoMbString(smoDatabaseProperties.MemoryUsedByMemoryOptimizedObjectsInKB),
+                                NumberOfUsers = smoDatabaseProperties.Users.Count.ToString(),
+                                Owner = smoDatabaseProperties.Owner.ToString(),
+                                SizeInMb = smoDatabaseProperties.Size.ToString("0.00") + " MB",
+                                SpaceAvailableInMb = ConvertKbtoMbString(smoDatabaseProperties.SpaceAvailable),
+                                Status = smoDatabaseProperties.Status.ToString()
+                            };
+                        }
 
                         // azure sql db doesn't have a sysadmin fixed role
                         var compatibilityLevelEnabled = !isDw &&
@@ -176,7 +197,7 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
         public override Task Save(DatabaseViewContext context, DatabaseInfo obj)
         {
             ConfigureDatabase(
-                context.Parameters.ConnectionUri,
+                context.Parameters,
                 obj,
                 context.Parameters.IsNewObject ? ConfigAction.Create : ConfigAction.Update,
                 RunType.RunNow);
@@ -186,39 +207,39 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
         public override Task<string> Script(DatabaseViewContext context, DatabaseInfo obj)
         {
             var script = ConfigureDatabase(
-                context.Parameters.ConnectionUri,
+                context.Parameters,
                 obj,
                 context.Parameters.IsNewObject ? ConfigAction.Create : ConfigAction.Update,
                 RunType.ScriptToWindow);
             return Task.FromResult(script);
         }
 
-        private CDataContainer CreateDatabaseDataContainer(string connectionUri, ConfigAction configAction, DatabaseInfo? database = null)
+        private CDataContainer CreateDatabaseDataContainer(InitializeViewRequestParams requestParams, ConfigAction configAction, DatabaseInfo? database = null)
         {
-            ConnectionInfo connectionInfo = this.GetConnectionInfo(connectionUri);
+            ConnectionInfo connectionInfo = this.GetConnectionInfo(requestParams.ConnectionUri);
             CDataContainer dataContainer = CDataContainer.CreateDataContainer(connectionInfo, databaseExists: configAction != ConfigAction.Create);
             if (dataContainer.Server == null)
             {
                 throw new InvalidOperationException(serverNotExistsError);
             }
-            string objectUrn = (configAction != ConfigAction.Create && database != null)
+            string objectUrn = requestParams.IsNewObject ? ((configAction != ConfigAction.Create && database != null)
                 ? string.Format(System.Globalization.CultureInfo.InvariantCulture,
                     "Server/Database[@Name='{0}']",
                     Urn.EscapeString(database.Name))
                 : string.Format(System.Globalization.CultureInfo.InvariantCulture,
-                    "Server");
-            dataContainer.SqlDialogSubject = dataContainer.Server.GetSmoObject(objectUrn);
+                    "Server")) : requestParams.ObjectUrn;
+            dataContainer.SqlDialogSubject = dataContainer.Server.GetSmoObject(objectUrn) as Database;
             return dataContainer;
         }
 
-        private string ConfigureDatabase(string connectionUri, DatabaseInfo database, ConfigAction configAction, RunType runType)
+        private string ConfigureDatabase(InitializeViewRequestParams requestParams, DatabaseInfo database, ConfigAction configAction, RunType runType)
         {
             if (database.Name == null)
             {
                 throw new ArgumentException("Database name not provided.");
             }
 
-            using (var dataContainer = CreateDatabaseDataContainer(connectionUri, configAction, database))
+            using (var dataContainer = CreateDatabaseDataContainer(requestParams, configAction, database))
             {
                 if (dataContainer.Server == null)
                 {
@@ -581,6 +602,16 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
 
             // previous loop did not find the prototype compatibility level in this server's compatability options, so treat compatibility levels as unsupported for this server
             return Array.Empty<string>();
+        }
+
+        /// <summary>
+        /// Converts value in KBs to MBs with two decimal places
+        /// </summary>
+        /// <param name="valueInKb"></param>
+        /// <returns>Returns as String</returns>
+        private string ConvertKbtoMbString(double valueInKb)
+        {
+            return (Math.Round(valueInKb / 1000, 2)).ToString("0.00") + " MB";
         }
     }
 }
