+++ src/Microsoft.SqlTools.ServiceLayer/ObjectManagement/ObjectTypes/Database/DatabaseInfo.cs
@@ -15,5 +15,14 @@ namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
         public string? RecoveryModel { get; set; }
         public string? CompatibilityLevel { get; set; }
         public string? ContainmentType { get; set; }
+        public string? DateCreated { get; set; }
+        public string? LastDatabaseBackup { get; set; }
+        public string? LastDatabaseLogBackup { get; set; }
+        public string? MemoryAllocatedToMemoryOptimizedObjectsInMb { get; set; }
+        public string? MemoryUsedByMemoryOptimizedObjectsInMb { get; set; }
+        public string? NumberOfUsers { get; set; }
+        public string? SizeInMb { get; set; }
+        public string? SpaceAvailableInMb { get; set; }
+        public string? Status { get; set; }
     }
 }
