--- src/Microsoft.SqlTools.ServiceLayer/ObjectManagement/ObjectTypes/DatabaseProperties/DatabasePropertiesViewContext.cs
@@ -1,20 +1,0 @@
-﻿//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-//
-
-using Microsoft.SqlTools.ServiceLayer.ObjectManagement.Contracts;
-
-namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
-{
-    public class DatabasePropertiesViewContext : SqlObjectViewContext
-    {
-        public DatabasePropertiesViewContext(InitializeViewRequestParams parameters) : base(parameters)
-        {
-        }
-
-        public override void Dispose()
-        {
-        }
-    }
-}
