--- src/Microsoft.SqlTools.ServiceLayer/ObjectManagement/ObjectTypes/DatabaseProperties/DatabasePropertiesViewInfo.cs
@@ -1,12 +1,0 @@
-﻿
-//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-//
-
-namespace Microsoft.SqlTools.ServiceLayer.ObjectManagement
-{
-    public class DatabasePropertiesViewInfo : SqlObjectViewInfo
-    {
-    }
-}
